// src/generated/ResourceAddress.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ResourceAddress = { "type": "root" } | { "type": "etcd", keyBase64: string, } | { "type": "etcdPrefix", prefixBase64: string, } | { "type": "zookeeper", path: string, } | { "type": "nacosConfig", group: string, dataId: string, };