// src/generated/WatchEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ResourceAddress } from "./ResourceAddress";
import type { WatchChangeKind } from "./WatchChangeKind";
import type { WatchStatusState } from "./WatchStatusState";

export type WatchEvent = { "kind": "status", subscriptionId: string, state: WatchStatusState, message: string | null, retryInMs: number | null, } | { "kind": "change", subscriptionId: string, change: WatchChangeKind, address: ResourceAddress, version: string | null, };